Only owners of the image can edit and delete the image
Allow only owners of the image to edit and delete the image.\n\t-Earlier all users were able to edit and delete any image in the application.\n\t-Changed the business logic to allow only owners of the image to edit and delete the respective image

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,9 +92,16 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
+        User user = (User)session.getAttribute("loggeduser");
 
+        if(image.getUser().getId() != user.getId()){
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError", "Only the owner of the image can edit the image");
+            return "images/image";
+        }
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
@@ -140,7 +147,16 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        User user = (User)session.getAttribute("loggeduser");
+
+        if(image.getUser().getId() != user.getId()){
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError", "Only the owner of the image can delete the image");
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
